Extract controller
