+++ src/main/java/com/jhuoose/todoose/Server.java
@@ -1,11 +1,14 @@
 package com.jhuoose.todoose;
 
+import com.jhuoose.todoose.controllers.ItemsController;
 import com.jhuoose.todoose.repositories.ItemNotFoundException;
 import com.jhuoose.todoose.repositories.ItemsRepository;
 import io.javalin.Javalin;
 
 import java.sql.DriverManager;
 import java.sql.SQLException;
+
+import static io.javalin.apibuilder.ApiBuilder.*;
 
 public class Server {
     public static void main(String[] args) throws SQLException {
@@ -14,25 +17,21 @@ public class Server {
         });
         var connection = DriverManager.getConnection("jdbc:sqlite:todoose.db");
         var itemsRepository = new ItemsRepository(connection);
+        var itemsController = new ItemsController(itemsRepository);
         app.events(event -> {
             event.serverStopped(() -> {
                 connection.close();
             });
         });
-        app.get("/items", ctx -> { ctx.json(itemsRepository.getAll()); });
-        app.post("/items", ctx -> {
-            itemsRepository.create();
-            ctx.status(201);
-        });
-        app.delete("/items/:identifier", ctx -> {
-            itemsRepository.delete(itemsRepository.getOne(Integer.parseInt(ctx.pathParam("identifier"))));
-            ctx.status(204);
-        });
-        app.put("/items/:identifier", ctx -> {
-            var item = itemsRepository.getOne(Integer.parseInt(ctx.pathParam("identifier")));
-            item.setDescription(ctx.formParam("description", ""));
-            itemsRepository.update(item);
-            ctx.status(204);
+        app.routes(() -> {
+            path("items", () -> {
+                get(itemsController::getAll);
+                post(itemsController::create);
+                path(":identifier", () -> {
+                    delete(itemsController::delete);
+                    put(itemsController::update);
+                });
+            });
         });
         app.exception(ItemNotFoundException.class, (e, ctx) -> { ctx.status(404); });
         app.start(System.getenv("PORT") == null ? 7000 : Integer.parseInt(System.getenv("PORT")));
